readyup|refactor: Regroup remote-fetch modules under a remote directory
Groups readyup's remote-fetch modules and their tests under one directory, with the test helpers they share moved to where coverage measures them. Behavior is unchanged.
