+++ packages/readyup/src/list/listCommand.ts
@@ -7,7 +7,6 @@ import { configError, kitLoadError, usageError } from '../errors.ts';
 import { EXIT_OK } from '../exitCodes.ts';
 import { KITS_DIR, resolveHomeDir } from '../kitsDir.ts';
 import { DEFAULT_CONFIG, loadConfig } from '../loadConfig.ts';
-import { loadRemoteManifest, RemoteManifestNotFoundError } from '../loadRemoteManifest.ts';
 import { DEFAULT_MANIFEST_PATH } from '../manifest/manifestPath.ts';
 import type { RdyManifest, RdyManifestKit } from '../manifest/manifestSchema.ts';
 import { ManifestNotFoundError, readManifest } from '../manifest/readManifest.ts';
@@ -15,8 +14,9 @@ import { discoverKitPackages } from '../packages/discoverKitPackages.ts';
 import { expandConfiguredPackages, type PackageKit } from '../packages/expandConfiguredPackages.ts';
 import type { DirectorySource, GlobalSource, LocalSource, NpmSource } from '../parseFromValue.ts';
 import { parseFromValue } from '../parseFromValue.ts';
-import { resolveBitbucketToken } from '../resolveBitbucketToken.ts';
-import { resolveGitHubToken } from '../resolveGitHubToken.ts';
+import { loadRemoteManifest, RemoteManifestNotFoundError } from '../remote/loadRemoteManifest.ts';
+import { resolveBitbucketToken } from '../remote/resolveBitbucketToken.ts';
+import { resolveGitHubToken } from '../remote/resolveGitHubToken.ts';
 import { resolvePackageRoot } from '../resolvePackageRoot.ts';
 import type { JsonListKitEntry, JsonListOutput } from '../schemas/index.ts';
 import { SCHEMA_VERSION } from '../schemas/listOutputSchema.ts';
